@@ -14,12 +14,12 @@
         }
     ],
     "require": {
-        "php": "^7.0"
+        "php": ">=7.1.3"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.0",
-        "mockery/mockery": "^0.9",
-        "phpunit/phpunit": "^6.0"
+        "mockery/mockery": "^1.0",
+        "phpunit/phpunit": "^7.0"
     },
     "autoload": {
         "psr-4": { "Werxe\\Skeleton\\": "src/" }
@@ -39,7 +39,8 @@
         }
     },
     "scripts": {
-        "cs": "vendor/bin/php-cs-fixer fix --verbose --ansi --show-progress=estimating",
+        "cs:check": "vendor/bin/php-cs-fixer fix --verbose --ansi --show-progress=estimating --diff --dry-run",
+        "cs:fix": "vendor/bin/php-cs-fixer fix --verbose --ansi --show-progress=estimating",
         "test": "vendor/bin/phpunit --verbose --colors=always"
     },
     "minimum-stability": "stable"
